Tests/LibWeb/Text/input/css/asf-disallowed-in-property-initial-value.html

<!doctype html>
<style>
    @property --foo {
        syntax: "*";
        inherits: true;
        initial-value: var(foo);
    }

    @property --bar {
        syntax: "*";
        inherits: true;
        initial-value: var(!this-does-not-match-argument-grammar);
    }
</style>
<div id="target"></div>
<script src="../include.js"></script>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
        println(document.styleSheets[0].cssRules[1].cssText);
    });
</script>
