Tests/LibWeb/Text/expected/css/asf-disallowed-in-property-initial-value.txt

@property --foo { syntax: "*"; inherits: true; }
@property --bar { syntax: "*"; inherits: true; }
